Use passlib for token hashing
We should just be using a well maintained upstream library for
this, rather than writing our own. This isn't too large a
dependency now because of wheels - so is ok.

This isn't enough by itself - we also need to deal with backwards
compatibility. And I need to understand what 'prefixes' are being
used for and why we need them.